move getOrder to api helper folder
