+++ app/pages/preview/controller.test.js
@@ -1,10 +1,7 @@
 import { getData } from 'buying-catalogue-library';
 import { orderApiUrl } from '../../config';
 import { logger } from '../../logger';
-import {
-  getOrder,
-  getPreviewPageContext,
-} from './controller';
+import { getPreviewPageContext } from './controller';
 import * as contextCreator from './contextCreator';
 import * as createServiceRecipientsDict from './helpers/createServiceRecipientsDict';
 import * as transformOrderItems from './helpers/transformOrderItems';
@@ -24,27 +21,6 @@ jest.mock('./helpers/transformOrderItems', () => ({
 }));
 
 describe('order summary preview controller', () => {
-  describe('getOrder', () => {
-    afterEach(() => {
-      jest.resetAllMocks();
-    });
-
-    const accessToken = 'access_token';
-    const orderId = 'order-id';
-
-    it('should call getData with the correct params', async () => {
-      getData.mockResolvedValueOnce({});
-
-      await getOrder({ orderId, accessToken });
-      expect(getData.mock.calls.length).toEqual(1);
-      expect(getData).toHaveBeenCalledWith({
-        endpoint: `${orderApiUrl}/api/v1/orders/order-id`,
-        accessToken,
-        logger,
-      });
-    });
-  });
-
   describe('getPreviewPageContext', () => {
     afterEach(() => {
       contextCreator.getContext.mockReset();
